Add ComputeResources and Image to Command and JobEnvironment DTOs
- Wire them through converters, resolver, agent launchers, entities
- Deal with trickle of data type changes and other side effects as development moved along and earlier assumptions were re-done

+++ genie-web/src/main/java/com/netflix/genie/web/data/services/impl/jpa/queries/projections/JobRequestProjection.java
@@ -94,7 +94,7 @@ public interface JobRequestProjection extends JobCommonFieldsProjection, SetupFi
      *
      * @return The amount of memory the user requested for this job in MB as an Optional
      */
-    Optional<Integer> getRequestedMemory();
+    Optional<Long> getRequestedMemory();
 
     /**
      * Get the timeout (in seconds) requested by the user for this job.
